E-Commerce chatbot with aws using Astra db(cassandra) and langchain and openai/requirements.txt
langchain-astradb
langchain
langchain-openai
datasets
pypdf
python-dotenv
flask

-e.